third_party/blink/web_tests/platform/generic/fast/css/getComputedStyle/getComputedStyle-border-shorthand-expected.txt

Tests that border shorthands are computed properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS computedStyle.getPropertyValue('border-bottom') is '10px solid rgb(255, 0, 0)'
PASS computedStyle.getPropertyValue('border-bottom') is '320px solid rgb(0, 0, 255)'
PASS computedStyle.getPropertyValue('border-bottom') is '0px none rgb(0, 128, 0)'
PASS computedStyle.getPropertyValue('border-top') is '10px solid rgb(255, 0, 0)'
PASS computedStyle.getPropertyValue('border-top') is '320px solid rgb(0, 0, 255)'
PASS computedStyle.getPropertyValue('border-top') is '0px none rgb(0, 128, 0)'
PASS computedStyle.getPropertyValue('border-right') is '10px solid rgb(255, 0, 0)'
PASS computedStyle.getPropertyValue('border-right') is '320px solid rgb(0, 0, 255)'
PASS computedStyle.getPropertyValue('border-right') is '0px none rgb(0, 128, 0)'
PASS computedStyle.getPropertyValue('border-left') is '10px solid rgb(255, 0, 0)'
PASS computedStyle.getPropertyValue('border-left') is '320px solid rgb(0, 0, 255)'
PASS computedStyle.getPropertyValue('border-left') is '0px none rgb(0, 128, 0)'
PASS computedStyle.getPropertyValue('border') is '320px solid rgb(255, 0, 0)'
PASS computedStyle.getPropertyValue('border') is ''
PASS successfullyParsed is true

TEST COMPLETE
